requirements.txt
fastwsgi
flask
requests
paho-mqtt
gpiozero
pigpio
#rpi.gpio
